% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_conditional.R
\name{em_conditional}
\alias{em_conditional}
\title{Estimates parameters for a continuous outcome model conditioned on a previously
estimated ordinal probit model. This function implements an EM algorithm where
the latent continuous variable \eqn{Y_1^*} is treated as missing data.}
\usage{
em_conditional(
  formula,
  data,
  formula_cond,
  theta_cond,
  vcov_cond = NULL,
  theta2_init = NULL,
  se_calc = TRUE,
  max_iter = 100,
  tol = 1e-06
)
}
\arguments{
\item{formula}{A \code{formula} object for the continuous outcome (\eqn{Y_2}).}

\item{data}{A data frame containing the variables in the model.}

\item{formula_cond}{The \code{formula} used for the ordinal outcome model (\eqn{Y_1}).}

\item{theta_cond}{Vector of estimated parameters from the ordinal model (\eqn{\theta_1}).}

\item{vcov_cond}{Covariance matrix of \eqn{\theta_1}. Optional; only required if \code{se_calc = TRUE}.}

\item{theta2_init}{Optional numeric vector of initial values for theta2.}

\item{se_calc}{Logical; whether to compute standard errors for Model 2 parameters. Defaults to TRUE.}

\item{max_iter}{Maximum number of EM iterations. Defaults to 100.}

\item{tol}{Convergence tolerance.}
}
\value{
A list containing estimates and convergence info.
}
\description{
Estimates parameters for a continuous outcome model conditioned on a previously
estimated ordinal probit model. This function implements an EM algorithm where
the latent continuous variable \eqn{Y_1^*} is treated as missing data.
}
