% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_smle.R
\name{orm_smle}
\alias{orm_smle}
\title{Sieve Maximum Likelihood Estimator (SMLE) under two-phase sampling designs for ordinal outcomes}
\usage{
orm_smle(
  formula,
  data,
  Bbasis,
  x_name,
  family = "probit",
  theta_init = NULL,
  se_calc = TRUE,
  se_method = "forward",
  verbose = TRUE,
  max_iter = 500,
  tol = 1e-06,
  se_max_iter = 1000,
  se_tol = 1e-08
)
}
\arguments{
\item{formula}{Formula for the ordinal outcome Y.}

\item{data}{Data frame. NA allowed only in \code{x_name} for S=0 subjects.}

\item{Bbasis}{Basis matrix for the spline basis.}

\item{x_name}{Name of the expensive covariate X.}

\item{family}{Character value specifying the distribution family, which is one of the following: "logit", "probit"}

\item{theta_init}{Optional initial vector for theta (beta, cutpoints).}

\item{se_calc}{Logical; whether to compute standard errors. Defaults to TRUE.}

\item{se_method}{Character value specifying the method for SE estimation: "forward" uses second-order forward-difference Hessian (default) and "numDeriv" uses Richardson extrapolation via \code{numDeriv::hessian()}.}

\item{verbose}{Logical; if TRUE, prints convergence info after estimation. Defaults to FALSE.}

\item{max_iter}{Maximum number of EM iterations. Defaults to 500.}

\item{tol}{Convergence tolerance for the optimizer. Defaults to 1e-6.}

\item{se_max_iter}{Maximum number of EM iterations for finding optimal p_vl in SE estimation. Defaults to 1000.}

\item{se_tol}{Convergence tolerance for the optimizer in SE estimation. Defaults to 1e-8.}
}
\value{
A list containing estimates and convergence info.
}
\description{
Sieve Maximum Likelihood Estimator (SMLE) under two-phase sampling designs for ordinal outcomes
}
