% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml_probit.R
\name{acml_probit}
\alias{acml_probit}
\title{ACML under Probit Link}
\usage{
acml_probit(formula, data, pi_values, theta_init = NULL)
}
\arguments{
\item{formula}{Formula for the ordinal outcome Y.}

\item{data}{Data frame containing all variables in \code{formula}. No NA allowed.}

\item{pi_values}{Numeric vector of sampling probabilities for ODS.}

\item{theta_init}{Optional initial vector for theta (beta, cutpoints).}
}
\value{
A list with containing estimates.
}
\description{
ACML under Probit Link
}
