% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml_probit.R
\name{acml_probit}
\alias{acml_probit}
\title{ACML under probit}
\usage{
acml_probit(formula, data, pi_values)
}
\arguments{
\item{formula}{A model formula specifying the probit model.}

\item{data}{A data frame containing the variables in the formula.}

\item{pi_values}{Numeric vector of sampling probabilities for ODS.}
}
\value{
A list with parameter estimates.
}
\description{
ACML under probit
}
