% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_joint.R
\name{em_joint}
\alias{em_joint}
\title{EM Algorithm for Joint Modeling - assuming both X and Y2 are available only for subjects selected in Phase 2}
\usage{
em_joint(
  formula1,
  formula2,
  data,
  B_basis,
  x_name,
  max_iter = 500,
  tol = 1e-06
)
}
\arguments{
\item{formula1}{Formula for the ordinal outcome Y1 (probit).}

\item{formula2}{Formula for the continuous outcome Y2 (Model 2).}

\item{data}{Full dataset containing Phase 1 and Phase 2.}

\item{B_basis}{Basis matrix for B-splines of Z1.}

\item{x_name}{Name of the expensive covariate X.}

\item{max_iter}{Maximum number of EM iterations. Defaults to 500.}

\item{tol}{Convergence tolerance.}
}
\value{
A list containing estimates and convergence info.
}
\description{
EM Algorithm for Joint Modeling - assuming both X and Y2 are available only for subjects selected in Phase 2
}
